move output into swappable reporters
