+++ docker-compose.yml
@@ -5,6 +5,7 @@ crossdock:
         - omega
     environment:
       - WAIT_FOR=alpha,omega
+      - REPORT=list
 
       - AXIS_CLIENT=alpha,omega
       - AXIS_SERVER=alpha,omega
